refactor: 랭킹 관련 컴포넌트 및 모델 패키지 위치 변경 (`home.ranking` -> `ranking`)
