+++ mobile/composeApp/src/commonMain/kotlin/com/yagubogu/ui/ranking/model/RankingProfileItem.kt
@@ -1,4 +1,4 @@
-package com.yagubogu.ui.home.ranking.model
+package com.yagubogu.ui.ranking.model
 
 sealed interface RankingProfileItem {
     val memberId: Long
