+++ mobile/composeApp/src/commonMain/kotlin/com/yagubogu/ui/ranking/model/RankingType.kt
@@ -1,4 +1,4 @@
-package com.yagubogu.ui.home.ranking.model
+package com.yagubogu.ui.ranking.model
 
 enum class RankingType {
     CHECK_IN,
